Collisions are a lot smoother now, and using sf::Vector2i where I can. The PlayerGraphicsComponent will still cast it to a float for rendering purposes. but that's ok.

+++ PlayerInputComponent.hpp
@@ -16,7 +16,7 @@ public:
     PlayerInputComponent(Entity &entity, PlayerPhysicsComponent &physicsComponent);
 
 private:
-    sf::Vector2f proposedVector_;
+    sf::Vector2i proposedVector_;
     PlayerPhysicsComponent &physicsComponent_;
 };
